+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bugfix: Fixed a bug that caused moderation buttons not to appear in historical messages if they were in your own channel
 - Minor: Added delete message button to the usercard if you have mod rights
 - Minor: Added chatter count to viewer list
+- Minor: Added `reset` command to clear cache
 
 - Minor: Remove TwitchEmotes.com attribution and the open/copy options when right-clicking a Twitch Emote. (#2214, #3136)
 - Minor: Strip leading @ and trailing , from username in /user and /usercard commands. (#3143)